% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CityFit.R
\name{CityFit}
\alias{CityFit}
\title{Fit distributed lag model for one city.}
\usage{
CityFit(root, criterion, city, cause = "all", control_ratio = 15,
  lags = 14, storm_id = NA, arglag)
}
\arguments{
\item{root}{The local path where the NMMAPS data is located.}

\item{criterion}{A character string which could be "rain50", "wind18", etc.}

\item{city}{A character string giving the abbreviation of city's name.}

\item{cause}{A character string giving the death outcome used in the model.}

\item{control_ratio}{A integer vector giving the number of control days for
one storm day.}

\item{lags}{A integer vector giving number of lagged days for one storm day.}

\item{storm_id}{A character vector giving the storm ID of one specific storm
when you want to investigate the effect of this storm only.}
}
\value{
Prints the abbreviation of city's name and returns a list of
\code{glm()} model object and predicted results by \code{crosspred()}
function from \code{dlnm} package.
}
\description{
This function uses dataframe created by \code{CrossoverData} and fits a
distributed lag model for this city.
}
\examples{
\dontrun{
fit_detroit <- CityFit(root = "~/tmp/NMMAPS/", criterion = "rain50",
                       city = "det", cause = "accident",
                       arglag = list(fun = "integer"))

lag_knots <- logknots(14, 4)
fit_detroit <- CityFit(root = "~/tmp/NMMAPS/", criterion = "rain50",
                       city = "det", cause = "accident",
                       arglag = list(fun = "ns", knots = lag_knots))
}

}

