% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CityStorm.R
\name{CityStorm}
\alias{CityStorm}
\title{Merge hurricane exposure to death outcomes.}
\usage{
CityStorm(root = "~/tmp/NMMAPS/", criterion = c(), city = c())
}
\arguments{
\item{root}{The local path where the NMMAPS data is located.}

\item{criterion}{A character string which could be "rain50", "wind18", etc.}

\item{city}{A character string giving the abbreviation of city's name.}
}
\value{
This function return a time-series dataframe containing daily death
counts and information about storms between 1988 to 2005 in this city.
}
\description{
This function merges the two datasets created by readCity() function and
readStorm() function for one city by date.
}
\examples{
CityStorm(root = "~/tmp/NMMAPS/", criterion = "rain75", city = "miam")

}

