% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossoverData.R
\name{CrossoverData}
\alias{CrossoverData}
\title{Generate a case-crossover datasets.}
\usage{
CrossoverData(root = c(), criterion = c(), city = c(),
  control_ratio = 15, lags = 14, storm_id = NA)
}
\arguments{
\item{root}{The local path where the NMMAPS data is located.}

\item{criterion}{A character string which could be "rain50", "wind18", etc.}

\item{city}{A character string giving the abbreviation of city's name.}

\item{control_ratio}{A integer vector giving the number of control days for
one storm day.}

\item{lags}{A integer vector giving number of lagged days for one storm day.}

\item{storm_id}{A character vector giving the storm ID of one specific storm
when you want to investigate the effect of this storm only.}
}
\value{
This function returns a dataframe with first \code{l} (l is equal to
the value of \code{lags} argument which is 14 by defauls) rows of NA, and
\code{n} stratums datasets. \code{n} is the number of storms if default
\code{storm_id} is used, otherwise there is only one stratum. Each stratum
has one row of storm day, l rows of lagged days, and 15 (by default) rows of
matched control days.
}
\description{
Generate a data frame for one city with storm days, lagged days, and matched
control days (non-storm days). The data frame will be directly used for
\code{crossbasis()} function in the \code{dlnm} package.
}
\examples{
\dontrun{
CrossoverData(root = "~/tmp/NMMAPS/", criterion = "rain75",
                       city = "miam",
                       storm_id = "Irene-1999")
}

}

