% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomhurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{geom_hurricane}
\usage{
geom_hurricane(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  scale_radii = 1,
  ...
)
}
\arguments{
\item{mapping}{default}

\item{data}{default}

\item{geom}{default}

\item{position}{default}

\item{show.legend}{default}

\item{inherit.aes}{default}

\item{scale_radii}{allows the user to plot wind radii charts with the radii scaled back to a certain percent of the maximum radii}

\item{...}{default}
}
\value{
a wind radii chart
}
\description{
This function returns a wind radii chart.
You should provide a dataset for a specific hurricane and a single observation time for that hurricane using.
}
\examples{
\dontrun{
get_map("Louisiana", zoom = 6, maptype = "toner-background") \%>\%
ggmap(extent = "device") +
geom_hurricane(data = storm_observation,
aes(x = longitude, y = latitude,
r_ne = ne, r_se = se, r_nw = nw, r_sw = sw,
fill = wind_speed, color = wind_speed))
}
}
