% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{Geom to produce hurricane radii graph}
\format{An object of class \code{GeomHurricane} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}) of length 4.}
\usage{
GeomHurricane
}
\description{
This Geom is meant to satisfy a Coursera assignment to build a custom Geom for
ggplot2. This is really only suited for the data source outlined in the function
in this package. The point of the graph is to show a radius of distance of a 
certain wind speed as recorded in the data for a given hurricane.
}
\keyword{datasets}
