% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane_geocode.R
\name{hurricane_geocode}
\alias{hurricane_geocode}
\title{Convert a storm center and wind radius into a set of arc points}
\usage{
hurricane_geocode(storm_data, x = "longitude", y = "latitude",
  r = "wind_radius", quadrant = "quadrant", wind_speed = "wind_speed",
  arcRes = 1)
}
\arguments{
\item{storm_data}{This input is a transformed set of data from the Colorado
State compilation of NHC data. The data must have been run through the 
\code{\link{tidy_tracks}} function. Alternatively, a different function can
be used so long as the input data contains all needed components.}

\item{x}{This should be the column name of the longitudinal coordinate of the
eye of the hurricanes.}

\item{y}{This should be the column name of the latitudinal coordinate of the
eye of the hurricanes.}

\item{r}{This is the name of the column that holds the farthest distance for
which the wind speed was recorded.}

\item{quadrant}{This is the name of the column containing the names of the 
spatial quadrants on a map. Currently only ne, se, sw, and nw are acceptable
entries.}

\item{wind_speed}{This is the name of the column containing the speeds of wind
for which measuments exist. There is an important implicit assumption that
faster winds will ALWAYS have a smaller \code{r} value than slower winds for 
a given \code{quadrant}. This fact is used to create arc point parings.}

\item{arcRes}{This specifies the number of angular degrees (as in degrees of
a circle not of a longitude nor latitude coord) between each point. Lower
numbers produce more points, and smoother curves when put on a map.}
}
\value{
A data.frame, tibble per \code{dplyr} package.
}
\description{
This function's primary purpose is to transform hurricane data
(longitude/latitude of eye and distance of wind speeds) into a set of points
outlining an arc for each provided wind speed while preserving the 
actual longitude and latitude of where the winds radius would reach.
The bulk of the work is done by functions from the \code{geosphere} package 
to code the distance to proper spacial coordinates. While generating points
with this function may be interesting, its true purpose is to support data
transforms necessary for \code{geom_hurricane}.
}
