% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixed_model.R
\name{Mixed_1}
\alias{Mixed_1}
\title{This function fits a mixed model with an random intercept for $community$
(i.e. $city$). Other covariates are indicator
variable for $year$ and $dow$.}
\usage{
Mixed_1(cb_to_mod, data_to_mod, cause)
}
\value{
Thins function returns a mixed model and the $crosspred$ object.
}
\description{
This function fits a mixed model with an random intercept for $community$
(i.e. $city$). Other covariates are indicator
variable for $year$ and $dow$.
}

