% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixed_model.R
\name{Mixed_bobyqa}
\alias{Mixed_bobyqa}
\title{This function fits a mixed model with an random intercept for $community$
(i.e. $city$). Other covariates are indicator
variable for $year$ and $dow$. The default optimizer is used in $glmer$.}
\usage{
Mixed_bobyqa(cb_to_mod, data_to_mod, cause)
}
\value{
Thins function returns a mixed model and the $crosspred$ object.
}
\description{
This function fits a mixed model with an random intercept for $community$
(i.e. $city$). Other covariates are indicator
variable for $year$ and $dow$. The default optimizer is used in $glmer$.
}
