% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaReg.R
\name{plot_meta}
\alias{plot_meta}
\title{Plot the prediction of a meta-regression.}
\usage{
plot_meta(meta_pred, title = c())
}
\arguments{
\item{meta_pred}{A \code{'crosspred'} class object.}

\item{title}{A character string giving the title of the plot.}
}
\description{
Plot the prediction of a meta-regression.
}

