% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaReg.R
\name{plot_meta_spl}
\alias{plot_meta_spl}
\title{Plot the prediction of a meta-regression for spline lag.}
\usage{
plot_meta_spl(meta_pred, title = c())
}
\arguments{
\item{meta_pred}{A \code{'crosspred'} class object.}

\item{title}{A character string giving the title of the plot.}
}
\description{
Plot the prediction of a meta-regression for spline lag.
}

