% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaReg.R
\name{pred_meta}
\alias{pred_meta}
\title{Prediction for a meta-regression}
\usage{
pred_meta(meta_model, exposure = c(), lags = 14, arglag)
}
\arguments{
\item{meta_model}{A model object of class \code{'mvmeta'}.}

\item{exposure}{A numeric vector giving the exposure value for prediction.}

\item{lags}{A integer vector giving number of lagged days for one storm day.}
}
\value{
Returns a list of class \code{'crosspred'}.
}
\description{
This function uses \code{crosspred()} function in the \code{dlnm} package to
predict the
}
\examples{
\dontrun{
exposure <- c(rep(0, 20), 1, rep(0, 20))
pred <- pred_meta(meta_model = meta, exposure = exposure,
                  arglag = list(fun = "ns", knots = logknots(14, 4)))
}

}

