% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pull_cities_char.R
\name{pull_cities_var}
\alias{pull_cities_var}
\title{Pull city-level characteristics}
\usage{
pull_cities_var(city_list)
}
\arguments{
\item{city_list}{A vector of abbreviation of cities' names.}
}
\value{
This function returns a dataframe giving city-level variables for all the cities in the \code{city_list}.
}
\description{
Pull city-level characteristics
}

