% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaReg.R
\name{pull_city_coef}
\alias{pull_city_coef}
\title{Pull Coefficients matrix.}
\usage{
pull_city_coef(CityFit)
}
\arguments{
\item{CityFit}{A fitted model object.}
}
\value{
A matrix of coefficients between the parameters estimates in the
model.
}
\description{
This function pulls coefficients matrix from \code{crosspred()}
object.
}

