% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaReg.R
\name{pull_city_vcov}
\alias{pull_city_vcov}
\title{Pull Variance-Covariance matrix.}
\usage{
pull_city_vcov(CityFit)
}
\arguments{
\item{CityFit}{A fitted model object.}
}
\value{
A matrix of variance-covariance between the parameters estimates in
the model.
}
\description{
This function pulls variance-covariance matrix from \code{crosspred()}
object.
}

