% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ext_tracks.R
\name{read_ext_tracks}
\alias{read_ext_tracks}
\title{Read in ext_tracks data from Colorado State}
\usage{
read_ext_tracks(file = "ebtrk_atlc_1988_2015.txt", ..., widths = NULL,
  colnames = NULL, degW = TRUE)
}
\arguments{
\item{file}{this is the filename (including path if necessary) to the txt file
containing the hurricane data.}

\item{...}{This can be used to pass alternative arguments to \code{readr::fwf()}
as seen in the following arguemnts.}

\item{widths}{This is a numeric vector of column widths for the fixed width
file. The default is set to the current convention.}

\item{colnames}{This is a character vector of column names we wish to apply to
the data being read in. The assumption is that there are no column names in 
the text file being read in. The default represents current convention.}

\item{degW}{This stands for "degrees West" and is a logical representing if the
text file data is coded such that the latitude is stated as "degrees West" and
thus be strictly positive. If TRUE (the default) the read in procedure will
transform the data by flipping the sign so that latitude is in universal
coordinates (like a number line). There is no corresponding argument for 
the longitude as hurricanes are only being track for the northern hemisphere
(always positive).}
}
\value{
data returned per specifications in readr::fwf. By default this is a
data.frame
}
\description{
This function is set to read in extended tracks hurricane data from Colorad
State NHC data compilation. This is simply a shell for \code{readr::read_fwf()}
with defaults set for the current data convenctions. Currently data is in a 
fixed with format and this function is dependent on that form.
}
