% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCity.R
\name{readCity}
\alias{readCity}
\title{Read NMMAPS data for each city.}
\usage{
readCity(root = "~/Documents/NMMAPS/", city = c(), collapseAge = TRUE,
  age_cat = NULL)
}
\arguments{
\item{root}{The local path where the NMMAPS data is located.}

\item{city}{A character string giving the abbreviation of city's name.}

\item{collapseAge}{Logic value specifying whether to aggregate records based
on age group.}

\item{age_cat}{A character string giving the selected age categories.}
}
\value{
This function returns a time-series dataframe containing date, death counts,
etc.
}
\description{
Read NMMAPS data for each city.
}
\examples{
readCity(root = "~/Documents/NMMAPS/", city = "miam")

}
