% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStorm.R
\name{readStorm}
\alias{readStorm}
\title{Read storm data.}
\usage{
readStorm(criterion = c(), city = c())
}
\arguments{
\item{criterion}{A character string which could be "rain50", "wind18", etc.}

\item{city}{A character string giving the abbreviation of city's name.}
}
\value{
This function returns a dataframe containing information of all
storms in the city between 1988 to 2005. Variables might be different
depending on the exposure crtierion.
}
\description{
Read storm data for one city based on exposure criterion. The storm data was
generated by "Hurricaneexposure" package and located in the subdirectory
"exposure".
}
\examples{
\dontrun{readStorm(criterion = "rain75", city = "miam")}

}
