% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-hurricane.R
\name{stat_hurricane}
\alias{stat_hurricane}
\title{Hurricane}
\usage{
stat_hurricane(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
\code{stat_hurricane} computes a specific data frame of coordinates required
by \link{geom_hurricane}. It contains the geographical points corresponding
to the four quarters of circle, specified by the origin point of the storm
and the wind radii in the four quadrants, for each wind speed level.
The \link[geosphere]{destPoint} function is adopted to generate the points
representing the quarters of circles.
}
\examples{
# When using `stat_hurricane`, you will typically need a sotrm
# observation data frame, with specific columns. You can find a
# test dataset along with this package named `ike`. Consider
# the following code as a test example:

\dontrun{
library(ggplot2)

ike_1 <- ike \%>\% dplyr::filter(latitude > 29 & latitude < 30)
plot <- ggplot(data = ike_1, mapping = aes(x = longitude, y = latitude)) +
stat_hurricane(mapping = aes(x = longitude, y = latitude, r_ne = ne,mr_se = se, r_sw = sw,
                             r_nw = nw, colour = wind_speed), alpha = 0) +
theme_minimal()
}

}
\seealso{
\link{geom_hurricane}
}
