% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomhurricane.R
\docType{data}
\name{StatHurricane}
\alias{StatHurricane}
\title{My approach was to create a new Stat, following the explanations in
\url{https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html}."
StatHurricane setup}
\format{
An object of class \code{StatHurricane} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatHurricane
}
\description{
My approach was to create a new Stat, following the explanations in
\url{https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html}."
StatHurricane setup
}
\keyword{datasets}
