% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormobservation.R
\name{storm_observation}
\alias{storm_observation}
\title{storm_observation}
\usage{
storm_observation(
  data,
  hurricane_id = "Katrina-2005",
  date_time = "2005-08-29 12:00:00"
)
}
\arguments{
\item{data}{must use stormdata}

\item{hurricane_id}{a string, Hurricane-year. The first character uppercase.}

\item{date_time}{a string, in yyyy-mm-dd hh:mm:ss format}
}
\value{
a data frame
}
\description{
Taking the dataset `stormdata`, this function creates a subset of a specific hurricane and a single observation time for that hurricane.
}
\examples{
\dontrun{
storm_observation("Ike-2008","2008-09-13 12:00:00")
storm_observation("Katrina-2005","2005-08-29 12:00:00")}


}
