% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormobservation.R
\docType{data}
\name{stormdata}
\alias{stormdata}
\title{Data on all storms in the Atlantic basin from 1988 to 2015.}
\format{
A data frame with 35 472 rows and 9 variables
\describe{
  \item{storm_id}{name of the hurricane and year (e.g. Katrina-2005)}
  \item{date}{date and time of the record (e.g. 2008-09-13 12:00:00)}
  \item{latitude}{latitudeof the center of the hurricane}
  \item{longitude}{longitude of the center of the hurricane}
  \item{wind_speed}{intensity of the wind in knots (34, 50, 60)}
  \item{ne}{distance (in meters) from the center of the storm of any location that experienced x-knot winds in the Northeast quadrant}
  \item{se}{distance (in meters) from the center of the storm of any location that experienced x-knot winds in the Southeast quadrant}
  \item{sw}{distance (in meters) from the center of the storm of any location that experienced x-knot winds in the Southwest quadrant}
  \item{nw}{distance (in meters) from the center of the storm of any location that experienced x-knot winds in the Northwest quadrant}
}
}
\source{
\url{<http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset/>}
}
\usage{
data("stormdata")
}
\description{
The dataset is based on the Extended Best Tract dataset and it was manipulated for the assignment "Build a New Geom" - part of Johns Hopkins University & Cousera Course "Building Data Visualization Tools".
}
\keyword{datasets}
