% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_tracks.R
\name{tidy_tracks}
\alias{tidy_tracks}
\title{Creates "tidy" data (long format) for hurricane data.}
\usage{
tidy_tracks(ext_tracks, makeDate = TRUE, keepcols = c("storm_name", "date",
  "longitude", "latitude"))
}
\arguments{
\item{ext_tracks}{This is data as read in via the \code{\link{read_ext_tracks}}
function.}

\item{makeDate}{This sets whether or not we wish to have the date columns
combined into a single date column.}

\item{keepcols}{This is a character vector of the columns from the original data
that are desired to be reatined. The defaults represent the minimum items
to satifsfy the assignment.}
}
\value{
A tidy data.frame, tibble per \code{dplyr} package
}
\description{
This function is used to "tidy" the dataset read in by the 
\code{\link{read_ext_tracks}} function. Although the assignment this function
is for requested to restate the data such that each quadrant had its own column
I decided this is just an annoying intermediate step when the ultimate goal is
to print the data which would require me to re-tidy the data.
}
