% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{Hurricane}
\usage{
geom_hurricane(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
\code{geom_hurricane} is a special implementation of \link[ggplot2]{geom_polygon}.
It is developed to draw the regions of a storm observation corresponding
to the same wind speed level, for the four quadrants defined by an origin
point. The contour of the "iso-speed" levels are connected by a `polygon`
grid object.
}
\section{Aestethics}{

\code{geom_hurricane} understands the following aesthetics (required in bold):
\itemize{
  \item \strong{x}: longitude
  \item \strong{y}: latitude
  \item \strong{r_ne}: wind radius for the NE quadrant, in nautical miles
  \item \strong{r_se}: wind radius for the SE quadrant, in nautical miles
  \item \strong{r_nw}: wind radius for the SW quadrant, in nautical miles
  \item \strong{r_sw}: wind radius for the NW quadrant, in nautical miles
  \item colour: (see \link[ggplot2]{geom_polygon})
  \item fill: (see \link[ggplot2]{geom_polygon})
  \item size: (see \link[ggplot2]{geom_polygon})
  \item linetype: (see \link[ggplot2]{geom_polygon})
  \item alpha: (see \link[ggplot2]{geom_polygon})
  \item scale_radii: scaling factor of the wind radius dimensions
}
}

\examples{
# When using `geom_hurricane`, you will typically need a storm
# observation data frame, with specific columns. You can find a
# test dataset along with this package named `ike`. Consider
# the following code as a test example:

\dontrun{
library(ggplot2)

ike_1 <- ike \%>\% dplyr::filter(latitude > 29 & latitude < 30)
map_plot <- ggmap::get_map("Lousiana", zoom = 6, maptype = "toner-background", source = "stamen")
map <- map_plot \%>\% ggmap::ggmap(extent = "device") +
    geom_hurricane(data = ike_1, mapping = aes(x = longitude, y = latitude, r_ne = ne, r_se = se, r_sw = sw,
                                               r_nw = nw, fill = wind_speed, colour = wind_speed, scale_radii = 0.4)) +
    scale_color_manual(name = "Wind speed (kts)", values = c("red", "orange", "yellow")) +
    scale_fill_manual(name = "Wind speed (kts)", values = c("red", "orange", "yellow"))
}

}
\seealso{
\link[ggplot2]{geom_polygon} for generic polygon
}
