% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Join_matched_cb.R
\name{Join_matched_cb}
\alias{Join_matched_cb}
\title{This function combine $matched crossbasis$.}
\usage{
Join_matched_cb(citylist, criterion, control_ratio, lags, arglag)
}
\description{
This function combine $matched crossbasis$.
}

