% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match_cb.R
\name{Match_cb}
\alias{Match_cb}
\title{Select matched observations from original $crossbasis$ matrix.}
\usage{
Match_cb(root = "~/tmp/NMMAPS/", criterion, city, control_ratio = 10,
  lags = 7, arglag = list(fun = "integer"))
}
\arguments{
\item{root}{The local path where the NMMAPS data is located.}

\item{criterion}{A character string which could be "rain50", "wind18", etc.}

\item{city}{A character string giving the abbreviation of city's name.}

\item{control_ratio}{A integer vector giving the number of control days for
one storm day.}

\item{lags}{A integer vector giving number of lagged days for one storm day.}
}
\value{
This function returns $crossbasis$ object.
}
\description{
Select matched observations from original $crossbasis$ matrix.
}

