% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaReg.R
\name{pull_city_coef}
\alias{pull_city_coef}
\title{Pull Coefficients and Variance-Covariance matrix.}
\usage{
pull_city_coef(CityFit)
}
\arguments{
\item{CityFit}{A fitted model object.}
}
\value{
A matrix of coefficients/variance-covariance between the parameters
estimates in the model.
}
\description{
This function pulls coefficients and variance-covariance matrix from \code{crosspred()}
object.
}

