% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaReg.R
\name{UniMetaReg}
\alias{UniMetaReg}
\title{Univariate Meta-regression.}
\usage{
UniMetaReg(city_list = c(), criterion = c(), cause = "all",
  method = "reml")
}
\arguments{
\item{city_list}{A list of abbreviation of cities' names}

\item{criterion}{A character string which could be "rain50", "wind18", etc.}

\item{cause}{A character vector giving the death outcome used in the model.}

\item{method}{ estimation method: \code{"fixed"} for fixed-effects models, \code{"ml"} or \code{"reml"} for random-effects models fitted through (restricted) maximum likelihood, \code{"mm"} for random-effects models fitted through method of moments, and \code{"vc"} for random-effects models fitted through variance components. See Details below. If \code{"model.frame"}, the model frame is returned, as in \code{\link{lm}} or \code{\link{glm}}.}
}
\value{
Returns a list of class \code{'mvmeta'}.
}
\description{
This function performs a univariate meta-regression using \code{mvmeta()}
function in the \code{mvmeta} package.
}
\examples{
\dontrun{
city_list <- c("no", "lkch", "miam", "jckv", "mobi")
meta <- UniMetaReg(city_list = city_list, criterion = "rain75",
                   cause = "accident")
}

}
\references{
Gasparrini A., Armstrong, B., Kenward M. G. (2012). Multivariate
meta-analysis for non-linear and other multi-parameter associations.
Statistics in Medicine. 31(29):3821-3839.
}

